Updated Controllers

+++ src/main/java/com/example/LMS_Backend/controller/CategoryController.java
@@ -2,8 +2,12 @@ package com.example.LMS_Backend.controller;
 
 import com.example.LMS_Backend.model.Category;
 import com.example.LMS_Backend.service.CategoryService;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.util.List;
 
@@ -15,34 +19,60 @@ public class CategoryController {
     @Autowired
     private CategoryService categoryService;
 
-    // Get all categories
-    @GetMapping
-    public List<Category> getAllCategories() {
-        return categoryService.getAll();
-    }
+   @RequestMapping(value={"/", "/list"}, method = RequestMethod.GET)
+   public String showCategoriesPage(Model model){
+       model.addAttribute("categories", categoryService.getAll());
+       return "/category/list";
+   }
 
-    // Get a category by ID
-    @GetMapping("/{id}")
-    public Category getCategoryById(@PathVariable Long id) {
-        return categoryService.get(id);
-    }
+   @RequestMapping(value="/add", method = RequestMethod.GET)
+    public String addCategoriesPage(Model model){
+       model.addAttribute("category", new Category());
+       return "/category/form";
+   }
 
-    // Add new category
-    @PostMapping
-    public Category addCategory(@RequestBody Category category) {
-        return categoryService.addNew(category);
-    }
+   @RequestMapping(value="/edit/{id}", method=RequestMethod.GET)
+    public String editCategoryPage(@PathVariable(name="id")Long id, Model model){
+       Category category = categoryService.get(id);
+       if(category != null){
+           model.addAttribute("category", category);
+           return "/category/form";
+       }
+       else{
+           return "redirect:/category/add";
+       }
+   }
 
-    // Update category
-    @PutMapping("/{id}")
-    public Category updateCategory(@PathVariable Long id, @RequestBody Category category) {
-        category.setId(id);
-        return categoryService.save(category);
-    }
 
-    // Delete category
-    @DeleteMapping("/{id}")
-    public void deleteCategory(@PathVariable Long id) {
-        categoryService.delete(id);
-    }
+   @RequestMapping(value="/save", method = RequestMethod.GET)
+    public String saveCategory(@Valid Category category, BindingResult bindingResult, final RedirectAttributes redirectAttributes){
+       if( bindingResult.hasErrors()) {
+           return "/category/form";
+       }
+           if( category.getId() == null){
+               categoryService.addNew(category);
+               redirectAttributes.addFlashAttribute("successMsg", "'" + category.getName() + "' category has been added.");
+               return "redirect:/category/add";
+           }
+           else{
+               Category updateCategory = categoryService.save(category);
+               redirectAttributes.addFlashAttribute("successMsg", "Changes for'" + category.getName() + "' has been done successfully.");
+               return "redirect:/category/add" + updateCategory.getId();
+           }
+   }
+
+   @RequestMapping(value="/remove/{id}", method = RequestMethod.GET)
+    public String removeCategory(@PathVariable(name="id")Long id, Model model){
+       Category category = categoryService.get(id);
+       if( category != null){
+           if(categoryService.hasUsage(category)){
+               model.addAttribute("categoryIssue", true);
+               return showCategoriesPage(model);
+           }
+           else{
+               categoryService.delete(id);
+           }
+       }
+       return "redirect:/categories/list";
+   }
 }
